+++ inst/examples/box_elements/accordion.R
@@ -2,26 +2,35 @@ accordion <- 'box(
   title = "Accordion Demo",
   width = NULL,
   accordion(
+    inputId = "accordion1",
     accordionItem(
-      id = 1,
-      title = "Accordion Item 1",
-      color = "danger",
+      title = "Accordion 1 Item 1",
+      color = "success",
       collapsed = TRUE,
       "This is some text!"
     ),
     accordionItem(
-      id = 2,
-      title = "Accordion Item 2",
-      color = "warning",
+      title = "Accordion 1 Item 2",
+      color = "success",
+      collapsed = FALSE,
+      "This is some text!"
+    )
+  ),
+  br(),
+  accordion(
+    inputId = "accordion2",
+    accordionItem(
+      title = "Accordion 2 Item 1",
+      color = "danger",
       collapsed = FALSE,
       "This is some text!"
     ),
     accordionItem(
-      id = 3,
-      title = "Accordion Item 3",
-      color = "info",
-      collapsed = FALSE,
+      title = "Accordion 2 Item 2",
+      color = "danger",
+      collapsed = TRUE,
       "This is some text!"
     )
   )
 )'
+